Fix nic device path check to ensure start_up_delay is used (bnc#777777)
Occasionally, after running ifup, there's some delay while the network
interface actually activates.  There was meant to be a sleep to
accommodate this, but it wasn't triggering due to incorrect string
interpolation in a "/sys/class/net/" file path.

Without the sleep, if the NIC doesn't come up in time, crowbar_join
will fail because the network is dead when it tries to apply the
resolver recipe.

Signed-off-by: Tim Serong <tserong@suse.com>