Create network-related data bags from proposal, not from template
We were creating the crowbar/admin_network and friends data bags in
create_proposal, which means that the network definitions came from the
template. Which is simply wrong since we want the network definitions
from the proposal.

Instead, create the data bags from apply_role_pre_chef_call (only if the
data bags don't exist yet, to avoid issues when re-applying the
proposal).