Only kill vlan interfaces if they are on the same parent interface
When we are about to create a vlan interface, we check that there's no
interface with the same vlan id, and if there is, we kill it.

However, we were not restricting this to the vlan interface on the same
parent interface. This might lead to situations where we want to use the
same vlan id on two interfaces, but crowbar doesn't allow this.