Fix the creation of the node.crowbar.bond_list attribute
Previously the node.crowbar.bond_list was maintained in barclamp_library.rb
(lookup_interface_info method) in a broken way. The names of the bond devices
it created where not aligned with the actual interfaces names on the system.

The first caller of lookup_interface_info would always get bond0, the next call
(with a differnt conduit) would get bond1, and so on.

The network cookbook however assigns the names in the order in which they are
returned from the node's crowbar.networks attribute (+ some sorting to make
sure that vlan and bridge interfaces are handled later).

This lead to broken installations of the admin node when a bastion network
is enabled, because Barclamp::Inventory.lookup_interface_info and the network
cookbook disagree about the device names.

To fix this this commit moves the maintenance of the bond_list attribute into
the network recipe and Barclamp::Inventory.lookup_interface_info will just use
what's in there.

https://bugzilla.novell.com/show_bug.cgi?id=833133