Epic refactor/shrink of the network barclamp to make it actually function.
It now no longer operates in demo mode, and can support creating new networks
on the fly.

Huge amounts of code that deal with trying to wrangle things all the way down to
the physical NIC layer on nodes is now gone -- it duplicated functionality that
was batter handled by the network recipe in any case.