Reduce local_port_range to not overlap with ports used by services
On systems with 1GBit or 10GBit interfaces we adjusted the sysctl value
net.ipv4.ip_local_port_range (to 1024-65000) to allow more concurrent outgoing
connections. This range however overlaps with a lot of ports that openstack
services are listening on. In the worst case it might happen that some services
do not start because the port they want to listen on is already used by some
other process for an outgoing connection. This patch reduces the local port
range again to a saner level.

fixes: https://bugzilla.novell.com/show_bug.cgi?id=839573