Allow lazy-defining of Nic in suse ifcfg template
There might be cases where a Nic is not available in the compile phase,
and so we can't pass it as a variable to the template.

For this use case, we allow passing a @nic_name variable that we use to
create the Nic object.

(cherry picked from commit 48eb801cf20dfc90758aac172d71f5bd966b454b)