Move SLE11 ovs workaround init scripts to network barclamp
This is the network barclamp's part of moving the SLE11 workaround init scripts
to configure the IP on ovs bridges during boot (formerly know as
ovs-usurp-config-br-*) from the neutron to the network barclamp. As the network
barclamp now handles the creation the bridges and create ifcfg files for them
the new init scripts could also be kept very simple (they just call ifup).